update add modal
